Refactored the whole project
This commit is way too big for my taste, but it turned out that I
started coding like a maniac at some point and lost track of my own
changes. I would be more careful if that was an open-source project that
I would be interested in releasing. I would rebuild the commits, but I
made those changes to use PPV[1] in Notion and synchronize with Google
Calendar, not to release a new project.

At some point, I was using Zappier to sync my Calendar with my pages in
Notion after I saw a tutorial on YouTube [2]. Pretty soon, I reached my
limit of tasks in Zappier, even with the paid plan -- mind you, when
you're a manager, you have lots of meetings.

There was also a problem with recurring tasks, and I saw another
tutorial on YouTube [2]. Still, I liked that idea less. Then I created a
concept of Schedules tasks, so every day it makes some pages in my
Notion.

The idea that there are so many lines of code here without tests makes
me very insecure, but so far, it's working, so which me luck.

[1]: https://www.youtube.com/watch?v=d93SGaf82OM&list=PLAl0gPKnL3V8s7dPXoo07mYnuErhWVk8b
[2]: https://www.youtube.com/watch?v=EFd6MQUuJzY
[3]: https://www.youtube.com/watch?v=uHXpIEM8HRs

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>