Remove "quickNote" field
I have never looked at this field and can't imagine using it for any
reason. I know this field is part of the standard template for PPV, but
I don't see its usefulness.

Yes, this open-source project advertises for a PPV integration with
Google Calendar, but I can't imagine someone else using it because I've
tailored it a lot for my needs. And there's no documentation whatsoever!

I should know if someone is using it soon because of this change.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>